+++ platform_unix.go
@@ -21,6 +21,19 @@ func registerLibFunc(fptr interface{}, handle uintptr, fname string) {
 	purego.RegisterLibFunc(fptr, handle, fname)
 }
 
+// tryRegisterLibFunc attempts to register a library function, returning an error if it fails
+// This is useful for optional functions that may not exist in all library builds
+func tryRegisterLibFunc(fptr interface{}, handle uintptr, fname string) error {
+	// First check if the symbol exists
+	_, err := purego.Dlsym(handle, fname)
+	if err != nil {
+		return err
+	}
+	// If it exists, register it
+	purego.RegisterLibFunc(fptr, handle, fname)
+	return nil
+}
+
 // getProcAddressPlatform gets the address of a symbol in a loaded library
 func getProcAddressPlatform(handle uintptr, name string) (uintptr, error) {
 	return purego.Dlsym(handle, name)